+++ src/app/shared/NoteContentKind30402.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import cx from "classnames"
   import {Tags} from "paravel"
   import {commaFormat} from "hurdak"
   import {Naddr} from "src/util/nostr"
@@ -6,11 +7,12 @@
   import Carousel from "src/partials/Carousel.svelte"
   import CurrencySymbol from "src/partials/CurrencySymbol.svelte"
   import Anchor from "src/partials/Anchor.svelte"
+  import Chip from "src/partials/Chip.svelte"
   import NoteContentLink from "src/app/shared/NoteContentLink.svelte"
   import NoteContentTopics from "src/app/shared/NoteContentTopics.svelte"
   import NoteContentKind1 from "src/app/shared/NoteContentKind1.svelte"
   import {router} from "src/app/router"
-  import {pubkey} from "src/engine"
+  import {pubkey, isDeleted} from "src/engine"
 
   export let note
   export let showMedia = false
@@ -20,6 +22,9 @@
   const {title, summary, location} = tags.getDict()
   const images = tags.type("image").values().all()
   const [price = 0, code = "SAT"] = tags.type("price").drop(1).first() || []
+  const address = Naddr.fromEvent(note).asTagValue()
+  const editLink = router.at("listings").of(address).at("edit").toString()
+  const deleteLink = router.at("listings").of(address).at("delete").toString()
 
   const sendMessage = () => {
     const initialMessage = `Hi, I'd like to make an offer on this listing:\n${Naddr.fromEvent(
@@ -28,12 +33,30 @@
 
     router.at("channels").of([$pubkey, note.pubkey]).cx({initialMessage}).push()
   }
+
+  $: deleted = $isDeleted(note)
 </script>
 
 <FlexColumn>
   <div class="flex flex-col gap-2">
     <div class="flex justify-between gap-2 text-xl">
-      <strong>{title}</strong>
+      <div class="flex items-center gap-3">
+        <strong class={cx({"line-through": deleted})}>
+          {title}
+        </strong>
+        {#if note.pubkey === $pubkey && !deleted}
+          <Anchor modal href={editLink} class="flex items-center">
+            <i class="fa fa-edit text-base text-lighter" />
+          </Anchor>
+          <Anchor modal href={deleteLink} class="flex items-center">
+            <i class="fa fa-trash text-base text-lighter" />
+          </Anchor>
+        {:else if deleted}
+          <Chip danger small>Deleted</Chip>
+        {:else}
+          <Chip small>Available</Chip>
+        {/if}
+      </div>
       <span class="whitespace-nowrap">
         <CurrencySymbol {code} />{commaFormat(price)}
         {code}
@@ -61,7 +84,9 @@
     </div>
   {/if}
   <NoteContentTopics {note} />
-  <div class="flex justify-center">
-    <Anchor button accent on:click={sendMessage}>Make an offer</Anchor>
-  </div>
+  {#if !deleted}
+    <div class="flex justify-center">
+      <Anchor button accent on:click={sendMessage}>Make an offer</Anchor>
+    </div>
+  {/if}
 </FlexColumn>