+++ src/app/views/ListingEdit.svelte
@@ -1,8 +1,10 @@
 <script lang="ts">
   import {onMount} from "svelte"
   import {Tags, now} from "paravel"
-  import {sleep} from "hurdak"
+  import {sleep} from 'hurdak'
+  import {whereEq} from 'ramda'
   import {Naddr, EventBuilder} from "src/util/nostr"
+  import {currencyOptions} from "src/util/i18n"
   import {secondsToDate, dateToSeconds} from "src/util/misc"
   import FlexColumn from "src/partials/FlexColumn.svelte"
   import Anchor from "src/partials/Anchor.svelte"
@@ -10,10 +12,12 @@
   import Field from "src/partials/Field.svelte"
   import Input from "src/partials/Input.svelte"
   import DateTimeInput from "src/partials/DateTimeInput.svelte"
+  import CurrencySymbol from "src/partials/CurrencySymbol.svelte"
+  import CurrencyInput from "src/partials/CurrencyInput.svelte"
   import ImageInput from "src/partials/ImageInput.svelte"
   import NoteImages from "src/app/shared/NoteImages.svelte"
   import Compose from "src/app/shared/Compose.svelte"
-  import {router} from "src/app/router"
+  import {router} from 'src/app/router'
   import {toastProgress} from "src/app/state"
   import {dereferenceNote, publishToZeroOrMoreGroups, getUserRelayUrls} from "src/engine"
 
@@ -23,17 +27,17 @@
   const onSubmit = async () => {
     const builder = EventBuilder.from(event)
 
-    builder.setTag("title", values.title)
-    builder.setTag("location", values.location)
-    builder.setTag("start", dateToSeconds(values.start).toString())
-    builder.setTag("end", dateToSeconds(values.end).toString())
+    builder.setTag('title', values.title)
+    builder.setTag('summary', values.summary)
+    builder.setTag('location', values.location)
+    builder.setTag('price', values.price.toString(), values.currency.code)
     builder.removeCircles()
     builder.setCreatedAt(now())
     builder.setContent(compose.parse())
-    builder.setImageMeta(images.getValue())
+    builder.setImages(images.getValue())
 
     const {pubs} = await publishToZeroOrMoreGroups(values.groups, builder.template, {
-      relays: getUserRelayUrls("write"),
+      relays: getUserRelayUrls('write'),
     })
 
     pubs[0].on("progress", toastProgress)
@@ -54,13 +58,15 @@
 
     if (event) {
       const tags = Tags.from(event)
+      const [price, code] = tags.type("price").first().slice(1)
 
       values = {
         groups: tags.circles().all(),
-        title: tags.getValue("title") || "",
-        location: tags.getValue("location") || "",
-        start: secondsToDate(tags.getValue("start") || now()),
-        end: secondsToDate(tags.getValue("end") || now()),
+        title: tags.getValue('title') || "",
+        summary: tags.getValue('summary') || "",
+        location: tags.getValue('location') || "",
+        price: parseInt(price || 0),
+        currency: currencyOptions.find(whereEq({code})) || currencyOptions.find(whereEq({code: "SAT"})),
       }
 
       // Wait for components to mount
@@ -68,8 +74,8 @@
 
       compose.write(event.content)
 
-      for (const image of tags.type("image").values().all()) {
-        images.addImage(new Tags([["url", image]]))
+      for (const image of tags.type('image').values().all()) {
+        images.addImage(new Tags([['url', image]]))
       }
     }
   })
@@ -83,25 +89,32 @@
       <Field label="Title">
         <Input bind:value={values.title} />
       </Field>
-      <div class="grid grid-cols-2 gap-2">
-        <div class="flex flex-col gap-1">
-          <strong>Start</strong>
-          <DateTimeInput bind:value={values.start} />
+      <Field label="Summary">
+        <Input bind:value={values.summary} />
+      </Field>
+      <Field label="Price">
+        <div class="grid grid-cols-3 gap-2">
+          <div class="col-span-2">
+            <Input type="number" placeholder="0" bind:value={values.price}>
+              <span slot="before">
+                <CurrencySymbol code={values.currency?.code || "SAT"} />
+              </span>
+            </Input>
+          </div>
+          <div class="relative">
+            <CurrencyInput bind:value={values.currency} />
+          </div>
         </div>
-        <div class="flex flex-col gap-1">
-          <strong>End</strong>
-          <DateTimeInput bind:value={values.end} />
-        </div>
-      </div>
+      </Field>
       <Field label="Location (optional)">
         <Input bind:value={values.location} />
       </Field>
       <Field label="Description">
-        <div class="rounded-xl border border-solid border-mid bg-white p-3 text-black">
+        <div class="rounded-xl border border-solid border-mid p-3 bg-white text-black">
           <Compose autofocus bind:this={compose} {onSubmit} />
         </div>
       </Field>
-      <NoteImages bind:this={images} bind:compose includeInContent />
+      <NoteImages bind:this={images} bind:compose />
       <div class="flex gap-2">
         <Anchor button tag="button" type="submit" class="flex-grow">Save</Anchor>
         <ImageInput