+++ src/partials/Chip.svelte
@@ -1,17 +1,21 @@
 <script lang="ts">
   import cx from "classnames"
 
-  export let theme = "dark"
+  export let dark = true
+  export let light = false
+  export let danger = false
+  export let warning = false
+  export let small = false
   export let onRemove = null
 
-  const className = cx(
-    $$props.class,
-    "inline-block rounded-full border border-solid py-1 px-2 cy-chip",
-    {
-      "border-lightest": theme === "dark",
-      "border-dark": theme === "light",
-    }
-  )
+  const className = cx($$props.class, "inline-block rounded-full border border-solid cy-chip", {
+    "border-dark": light,
+    "border-lightest": dark,
+    "!border-danger": danger,
+    "!border-warning": warning,
+    "py-1 px-2": !small,
+    "py-0.5 px-2 text-sm": small,
+  })
 </script>
 
 <div class={className} on:click>